Move wallet controls to nickname overlay
