+++ frontend/src/components/NicknameScreen.tsx
@@ -14,10 +14,18 @@ interface NicknameScreenProps {
   onBetChange: (value: string) => void
   onBetBlur: () => void
   balance: number
+  currentBet: number
   onStart: () => void
   startDisabled: boolean
   startDisabledHint?: string
   startLabel?: string
+  walletAddress?: string | null
+  walletSol?: number
+  walletUsd?: number | null
+  usdRate?: number | null
+  walletLoading?: boolean
+  onRefreshWallet?: () => void
+  onTopUp?: () => void
 }
 
 export function NicknameScreen({
@@ -32,16 +40,28 @@ export function NicknameScreen({
   onBetChange,
   onBetBlur,
   balance,
+  currentBet,
   onStart,
   startDisabled,
   startDisabledHint,
-  startLabel
+  startLabel,
+  walletAddress,
+  walletSol,
+  walletUsd,
+  usdRate,
+  walletLoading,
+  onRefreshWallet,
+  onTopUp
 }: NicknameScreenProps) {
   const handleSubmit = (event: FormEvent) => {
     event.preventDefault()
     if (startDisabled) return
     onStart()
   }
+
+  const formattedSol = typeof walletSol === 'number' ? walletSol.toFixed(3) : '0.000'
+  const formattedUsd = typeof walletUsd === 'number' ? walletUsd.toFixed(2) : '—'
+  const showWallet = Boolean(walletAddress)
 
   return (
     <div id="nicknameScreen" className={visible ? 'overlay' : 'overlay hidden'}>
@@ -90,6 +110,57 @@ export function NicknameScreen({
               ))}
             </div>
           </div>
+          <div className="account">
+            <div className="account-row">
+              <span className="account-label">Баланс</span>
+              <span className="account-value" id="balanceValue">
+                {formatNumber(balance)}
+              </span>
+            </div>
+            <div className="account-row">
+              <span className="account-label">Текущая ставка</span>
+              <span className="account-value">
+                {formatNumber(currentBet)}
+              </span>
+            </div>
+          </div>
+          {showWallet && (
+            <div className="wallet-section">
+              <div className="wallet-row">
+                <span className="wallet-label">SOL</span>
+                <span className="wallet-value">{formattedSol}</span>
+              </div>
+              <div className="wallet-row">
+                <span className="wallet-label">USD</span>
+                <span className="wallet-value">
+                  {formattedUsd}
+                  {usdRate ? <span className="wallet-rate">@{usdRate.toFixed(2)}</span> : null}
+                </span>
+              </div>
+              <div className="wallet-address" title={walletAddress ?? ''}>
+                <span className="wallet-label">Кошелек</span>
+                <span className="wallet-hash">{walletAddress}</span>
+              </div>
+              <div className="wallet-actions">
+                <button
+                  type="button"
+                  className="wallet-button"
+                  onClick={onTopUp}
+                  disabled={walletLoading}
+                >
+                  {walletLoading ? 'Обработка...' : 'Пополнить'}
+                </button>
+                <button
+                  type="button"
+                  className="wallet-button secondary"
+                  onClick={onRefreshWallet}
+                  disabled={walletLoading}
+                >
+                  {walletLoading ? 'Обновление...' : 'Обновить'}
+                </button>
+              </div>
+            </div>
+          )}
           <div className="bet-control">
             <label htmlFor="betInput">Ставка перед стартом</label>
             <input
